Fix integration test issues (#168)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
I tested the Humble SSC wrapper on Blue Lexus using the binaries I got
form Hexagon.
This was basically a simple smoke/regression test that just lane follows
on the innocation drive.
driver_shutdown_ros2 (hence assuming dirver_shutdown too) just needed to
be built for it to successfully launch.
I will try to test on Fusion the ROS1 package if I have time.

NOTE: I tested with all develop images which are in Foxy except the
ssc_wrapper.
Initially either the engage request was not being received or the ssc
was just not responding (it did not shift gear etc after the engage
service returned successfull).
This prompted me to use ros1_bridge and platform_ros2 as Humble, but
that required minimum of 11.7 Cuda to work (by platform_ros2). This made
me reinstall Nvidia and Kvaser drivers on the vehicle.

In the process, I also remembered to install libas-common on the host
(it is there in the image), which is requirement for Humble SSC binary
to work. I am guessing this made the trick, because now with the same
setup, the Lexus was able to engage and drive using trajectory_follower
control plugin.


<!--- Describe your changes in detail -->

## Related GitHub Issue
NA
<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
ARC-234
<!-- e.g. CAR-123 -->

## Motivation and Context
Humble Upgrade smoke test
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Innovation Drive Basic Lane Following
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [X] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] I have added any new packages to the sonar-scanner.properties file
- [X] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.