Use new ROS1 repo to build dataspeed controller messages (#169)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
We needed steering wheel data from ROS1 controllers. However, those
needed to be in ROS2 so that we have easier way to analyze. So in order
to add the ROS2 we needed to fork the ROS1 repo and add it. After that,
we needed to use the new repo to build otherwise the md5sum issue would
occur.

NOTE: I believe we are not releasing that forked repo, so not creating a
PR for that extra branch I made. Here is the new repo:
https://github.com/usdot-fhwa-stol/carma-dbw-mkz-ros.git
And this is the change from the 1.2.4 version we use to add the ROS2 for
the dbw_mkz_msgs

https://github.com/usdot-fhwa-stol/carma-dbw-mkz-ros/compare/1.2.4...1.2.4-ros2
<!--- Describe your changes in detail -->

## Related GitHub Issue
NA
<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
ARC-254
<!-- e.g. CAR-123 -->

## Motivation and Context
Control improvement scripts needed these data.
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Locally published messages in ROS2 and ROS1 and made sure they are
bridged
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [X] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.